Fix OpenMP errors on Windows

+++ libeasea/include/algorithms/moea/Casrea.h
@@ -103,7 +103,7 @@ typename Casrea<TIndividual, TRandom>::TPopulation Casrea<TIndividual, TRandom>:
 #ifdef USE_OPENMP
     EASEA_PRAGMA_OMP_PARALLEL
 #endif
-        for (size_t i = 0; i < offspring.size(); ++i)
+        for (int i = 0; i < offspring.size(); ++i)
         {
                 TI &child = offspring[i];
                 this->getMutation()(child);
