+++ libeasea/include/algorithms/moea/Cibea.h
@@ -165,7 +165,7 @@ typename Cibea<TIndividual, TRandom>::TPopulation Cibea<TIndividual, TRandom>::r
 #ifdef USE_OPENMP
     EASEA_PRAGMA_OMP_PARALLEL
 #endif
-        for (size_t i = 0; i < offspring.size(); ++i)
+        for (int i = 0; i < offspring.size(); ++i)
         {
                 TI &child = offspring[i];
                 this->getMutation()(child);
