+++ libeasea/include/algorithms/moea/CmoeaAlgorithm.h
@@ -51,7 +51,7 @@ CmoeaAlgorithm<TPopulation, TRandom>::CmoeaAlgorithm(TR random, TP &problem, con
 #ifdef USE_OPENMP
 EASEA_PRAGMA_OMP_PARALLEL
 #endif
-	for (size_t i = 0; i < initial.size(); ++i)
+	for (int i = 0; i < initial.size(); ++i)
         {
                 TI &individual = m_population[i];
                 individual.m_variable = initial[i];
